feat: 로또(Lotto) 클래스 추가
- 패키지 이동 (package lotto -> package lotto.model.lotto)
